backend/pyproject.toml
[tool.poetry]
name = "convertifyfm"
version = "0.1.0"
description = "Web app to manage your favourite music on Spotify."
authors = ["Andrew Borovikov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.8"
asgiref = "3.4.1"
certifi = "2021.5.30"
cffi = "1.14.6"
charset-normalizer = "2.0.6"
cryptography = "35.0.0"
defusedxml = "0.7.1"
deprecated = "1.2.13"
django = "3.2.7"
django-braces = "1.14.0"
django-cors-headers = "3.13.0"
django-oauth-toolkit = "1.5.0"
django-rest-framework-social-oauth2 = "1.1.0"
djangorestframework = "3.12.4"
idna = "3.2"
jwcrypto = "1.0"
oauthlib = "3.1.1"
pycparser = "2.20"
pyjwt = "2.1.0"
python3-openid = "3.2.0"
pytz = "2021.1"
requests = "2.26.0"
requests-oauthlib = "1.3.0"
six = "1.16.0"
social-auth-app-django = "5.0.0"
social-auth-core = "4.1.0"
sqlparse = "0.4.2"
urllib3 = "1.26.7"
wrapt = "1.12.1"
python-decouple = "^3.6"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
